third_party/blink/web_tests/virtual/disable-ua-client-hint/external/wpt/client-hints/sandbox/iframe-csp-same-origin.https-expected.txt

This is a testharness.js-based test.
FAIL CSP sandboxed iframe with same-origin flag does send client hint headers assert_equals: message from opened frame expected "PASS" but got "FAIL sec-ch-ua True None"
Harness: the test ran to completion.
